requirements.txt
asgiref==3.4.1
cffi==1.14.6
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2.6
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
jwt==1.2.0
Markdown==3.3.4
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
